Fixed Program Stop at unsolvable puzzle
-Solved! But I dont really know why. I removed the
IndexOutOfBoundsException on increment() and instead made it return null
in that case. This triggered setValues() in Puzzle to try to iterate
over a null object, which caused an exception I caught. Somehow
execution continued after that...I dunno why but it works.

-Note that I think program will still quit completely if a file was not
found, even if there may be more existing input files waiting to be
analyzed. But I'm going to bed :P