Made the hasDuplicates function more effiecent.
Instead of using a hashmap, I dumped all the values in the list into a
hashset via the hashset constructor. After the values were placed into
the hashset I then compared the sizes of the hashset and the original
list. If the hash set is smaller then the list then a duplicate exists
in the list. The boolean result is then returned.