Fixed Bugs/Commented/Cleaned
FileData.java has been renamed to SudokuFileReader.java.

A lot of cases that broke the file reader are now handled with a new
exception “SudokuFileReadException”.

The code has been heavily commented for ease of readability and should
make a lot more sense to someone who has never worked with the code
before.

Values that are in the test files that do not match the patterns of a
comment, puzzle variable or dimension variable will now throw an
exception.

Reformatted the UnitTests to match the new method structure. All the
unit test properly work as the have before.