File Reader and Puzzle.java have been fixed
Although the files have been fixed there are still errors in the other
files, these need to be looked over and fixed to handle a 2D array.

There is also a weird bug with the print function that is not printing
the value of the puzzle right, or the values are not being entered in
the array right.