Added Comments and Handled a unhandled exception
Re-Commented:
Main.java to explain and condense the code.
Cell.java to have java doc and explain the code. I also initialized the
box in the default constructor and had the others call the default via
the super() method.
Handled the FileNotFoundException if a file location was not found from
the hard coded input.