Added Comments Cleaned up Backtracking A little.
ArrayList has a constructor that takes a collection as a parameter so
you do not need to iterate over the tree set and place each one in the
array list one by one, just set a "new arraylist(tree set here)".
Commented some code in the functions and depricated all of the older
functions not being used.